requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
certifi==2021.5.30
chardet==4.0.0
discord.py==1.7.3
giphy-client==1.0.0
idna==3.2
multidict==5.1.0
python-dateutil==2.8.1
python-dotenv==0.17.1
six==1.16.0
typing-extensions==3.10.0.0
urllib3==1.26.5
yarl==1.6.3
